Refactoring Context objects into util dir
