+++ site/src/util/contexts.js
@@ -3,3 +3,4 @@ import { createContext } from "react";
 const LiftedContext = createContext({ lifted: false, setLifted: () => {} });
 
 export default LiftedContext;
+export { LiftedContext };
